Reservation added

+++ database/migrations/2024_04_04_012519_create_wallets_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('user_id') ->references('id')->on('users')->onDelete('cascade');
-            $table->float('balance');
+            $table->decimal('balance',10,2);
             $table->timestamps();
         });
     }
