+++ database/migrations/2024_04_04_151653_create_payments_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
             $table->enum('status',$allowed=['paid','unpaid','semi-paid']);
-            $table->double('amount');
+            $table->decimal('amount',10,2);
             $table->timestamps();
         });
     }
